requirements.txt
aniso8601==9.0.1
asttokens==2.2.1
attrs==22.1.0
backcall==0.2.0
bidict==0.22.0
black==22.12.0
certifi==2022.12.7
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
Fiona==1.8.22
Flask==2.2.2
Flask-RESTful==0.3.9
geopandas==0.12.1
importlib-metadata==5.1.0
ipykernel==6.17.1
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==7.4.8
jupyter_core==5.1.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
munch==2.5.0
mypy-extensions==0.4.3
nest-asyncio==1.5.6
numpy==1.23.5
packaging==21.3
pandas==1.5.2
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.4
prompt-toolkit==3.0.33
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
pyproj==3.4.0
python-dateutil==2.8.2
pytz==2022.6
pyzmq==24.0.1
Shapely==1.8.5.post1
six==1.16.0
stack-data==0.6.2
tomli==2.0.1
tornado==6.2
traitlets==5.6.0
typing_extensions==4.4.0
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.11.0